+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/StatesMachine.kt
@@ -1,24 +1,24 @@
 package dev.inmo.micro_utils.fsm.common
 
-import dev.inmo.micro_utils.coroutines.*
+import dev.inmo.micro_utils.coroutines.launchSafelyWithoutExceptions
+import dev.inmo.micro_utils.coroutines.subscribeSafelyWithoutExceptions
 import kotlinx.coroutines.*
-import kotlinx.coroutines.flow.asFlow
-
-private suspend fun <I : State> StatesMachine.launchStateHandling(
-    state: State,
-    handlers: List<StateHandlerHolder<out I>>
-): State? {
-    return handlers.firstOrNull { it.checkHandleable(state) } ?.run {
-        handleState(state)
-    }
-}
 
 /**
  * Default [StatesMachine] may [startChain] and use inside logic for handling [State]s. By default you may use
  * [DefaultStatesMachine] or build it with [dev.inmo.micro_utils.fsm.common.dsl.buildFSM]. Implementers MUST NOT start
  * handling until [start] method will be called
  */
-interface StatesMachine : StatesHandler<State> {
+interface StatesMachine<T : State> : StatesHandler<T, T> {
+    suspend fun launchStateHandling(
+        state: T,
+        handlers: List<CustomizableHandlerHolder<in T, T>>
+    ): T? {
+        return handlers.firstOrNull { it.checkHandleable(state) } ?.run {
+            handleState(state)
+        }
+    }
+
     /**
      * Starts handling of [State]s
      */
@@ -27,15 +27,15 @@ interface StatesMachine : StatesHandler<State> {
     /**
      * Start chain of [State]s witn [state]
      */
-    suspend fun startChain(state: State)
+    suspend fun startChain(state: T)
 
     companion object {
         /**
          * Creates [DefaultStatesMachine]
          */
-        operator fun invoke(
-            statesManager: StatesManager,
-            handlers: List<StateHandlerHolder<*>>
+        operator fun <T: State> invoke(
+            statesManager: StatesManager<T>,
+            handlers: List<CustomizableHandlerHolder<in T, T>>
         ) = DefaultStatesMachine(statesManager, handlers)
     }
 }
@@ -44,14 +44,14 @@ interface StatesMachine : StatesHandler<State> {
  * Default realization of [StatesMachine]. It uses [statesManager] for incapsulation of [State]s storing and contexts
  * resolving, and uses [launchStateHandling] for [State] handling
  */
-class DefaultStatesMachine (
-    private val statesManager: StatesManager,
-    private val handlers: List<StateHandlerHolder<*>>
-) : StatesMachine {
+class DefaultStatesMachine <T: State>(
+    private val statesManager: StatesManager<T>,
+    private val handlers: List<CustomizableHandlerHolder<in T, T>>
+) : StatesMachine<T> {
     /**
      * Will call [launchStateHandling] for state handling
      */
-    override suspend fun StatesMachine.handleState(state: State): State? = launchStateHandling(state, handlers)
+    override suspend fun StatesMachine<in T>.handleState(state: T): T? = launchStateHandling(state, handlers)
 
     /**
      * Launch handling of states. On [statesManager] [StatesManager.onStartChain],
@@ -60,7 +60,7 @@ class DefaultStatesMachine (
      * [StatesManager.endChain].
      */
     override fun start(scope: CoroutineScope): Job = scope.launchSafelyWithoutExceptions {
-        val statePerformer: suspend (State) -> Unit = { state: State ->
+        val statePerformer: suspend (T) -> Unit = { state: T ->
             val newState = launchStateHandling(state, handlers)
             if (newState != null) {
                 statesManager.update(state, newState)
@@ -83,7 +83,7 @@ class DefaultStatesMachine (
     /**
      * Just calls [StatesManager.startChain] of [statesManager]
      */
-    override suspend fun startChain(state: State) {
+    override suspend fun startChain(state: T) {
         statesManager.startChain(state)
     }
 }